moved legend interface to a dedicated class
